Я говорил что мне не нравится что есть люди которые просто так бьют культ первые?
Вот вам, заклинание проклятья, которое пусть снимается только т3 зельем
Заклинание проклятья крови
Если культиста ударят первым он проклянёт напавшего, после ему надо прочитать заклинание что бы перезарядить проклятье, если он ударит первым то проклятье уйдёт в кд, через стрелы и магию это не работает
